newUpdate

+++ Application/Response/TransportService/TransportServiceResponse.cs
@@ -13,9 +13,9 @@ namespace Application.Response.TransportService
         public string Name { get; set; }
         public TransportType TransportType { get; set; }
         public string Description { get; set; }
-        public double TransportPrice { get; set; }
-        public double PricePerKg { get; set; }
-        public double PricePerAmount { get; set; }
+        public decimal TransportPrice { get; set; }
+        public decimal PricePerKg { get; set; }
+        public decimal PricePerAmount { get; set; }
         public bool IsActive { get; set; }
     }
 }
