+++ Domain/Entity/TransportService.cs
@@ -12,9 +12,9 @@ namespace Domain.Entity
         public string Name { get; set; }
         public TransportType TransportType { get; set; }
         public string Description { get; set; }
-        public double TransportPrice { get; set; }
-        public double PricePerKg { get; set; }
-        public double PricePerAmount { get; set; }
+        public decimal TransportPrice { get; set; }
+        public decimal PricePerKg { get; set; }
+        public decimal PricePerAmount { get; set; }
         
         public bool IsActive { get; set; }
         //
